asdas

+++ Canvas2D/include/Vector2.h
@@ -52,7 +52,7 @@ public:
        return( aux );
    }
 
-   Vector2 operator * (const int s)
+   Vector2 operator * (const float s)
    {
        Vector2 aux(x * s, y * s);
        return(aux);
